requirements.txt
# Install PyTorch with CUDA first:
# pip install torch --index-url https://download.pytorch.org/whl/cu121

torch>=2.0
numpy>=1.24
networkx>=3.0
sympy>=1.12
pyyaml>=6.0
wandb>=0.15
tqdm>=4.65
requests>=2.31
streamlit>=1.28
plotly>=5.18.0
